pyproject.toml
[project]
name = "sql-security-investigator"
version = "1.0.0"
description = "OpenEnv SQL security investigation benchmark for the Meta PyTorch hackathon."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "pydantic",
    "openai",
    "requests",
    "httpx",
    "uvicorn",
    "python-dotenv",
    "openenv-core[core]",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
